+++ code/frontend/src/features/login/components/login-form.types.ts
@@ -1,0 +1,9 @@
+/**
+ * LoginForm component types
+ */
+export type LoginFormHandle = {
+  /**
+   * Call submit from outside
+   */
+  submit: () => void;
+};
