+++ code/frontend/src/shared/components/Footer.tsx
@@ -1,49 +1,29 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useCallback } from 'react';
 import './Footer.css';
+import type { FooterProps, FKey } from '../types/footer';
 
-
-// F1～F12のキー配列
-const F_KEYS = Array.from({ length: 12 }, (_, i) => `F${i + 1}`);
-
-
-/**
- * フッターコンポーネントのProps
- * - onFKeyPress: F1～F12ボタン押下時のコールバック群
- * - buttonNames: F1～F12ボタンの割り当て名
- */
-type FooterProps = {
-  /**
-   * F1～F12ボタン押下時のコールバック群。例: { F1: () => {}, F2: () => {}, ... }
-   */
-  onFKeyPress?: Partial<Record<`F${number}`, () => void>>;
-  /**
-   * F1～F12ボタンの割り当て名。例: { F1: 'ログイン', F2: 'キャンセル', ... }
-   */
-  buttonNames?: Partial<Record<`F${number}`, string>>;
-};
-
-
+// F1-F12 key array
+const F_KEYS: FKey[] = Array.from({ length: 12 }, (_, i) => `F${i + 1}` as FKey);
 
 /**
- * フッターコンポーネント
- * F1～F12キーやボタン押下でコールバックを実行
+ * Footer component
+ * Handles F1-F12 key/button press and executes callbacks
  */
 const Footer: React.FC<FooterProps> = ({ onFKeyPress, buttonNames }) => {
   useEffect(() => {
     /**
-     * F1～F12キー押下時の処理
+     * Handle F1-F12 keydown event
      */
     const handleKeyDown = (e: KeyboardEvent) => {
       if (e.key.startsWith('F') && !isNaN(Number(e.key.slice(1)))) {
         const idx = Number(e.key.slice(1));
         if (idx >= 1 && idx <= 12) {
-          e.preventDefault(); // ブラウザのデフォルト動作を抑止
+          e.preventDefault(); // Prevent browser default
           document.getElementById(`footer-btn-${e.key}`)?.classList.add('active');
           setTimeout(() => {
             document.getElementById(`footer-btn-${e.key}`)?.classList.remove('active');
           }, 150);
-          // コールバック呼び出し
-          onFKeyPress?.[e.key as keyof typeof onFKeyPress]?.();
+          onFKeyPress?.[e.key as FKey]?.();
         }
       }
     };
@@ -52,19 +32,18 @@ const Footer: React.FC<FooterProps> = ({ onFKeyPress, buttonNames }) => {
   }, [onFKeyPress]);
 
   /**
-   * ボタンクリック時の処理
-   * @param key F1～F12
+   * Handle button click
    */
-  const handleClick = (key: string) => {
-    onFKeyPress?.[key as keyof typeof onFKeyPress]?.();
-  };
+  const handleClick = useCallback((key: FKey) => {
+    onFKeyPress?.[key]?.();
+  }, [onFKeyPress]);
 
   return (
     <footer className="footer">
-      {/* F1～F12ボタンを動的に生成 */}
+      {/* Render F1-F12 buttons dynamically */}
       {F_KEYS.map((key) => {
-        const name = buttonNames?.[key as keyof typeof buttonNames];
-        if (!name) return null; // 割り当て名がない場合はボタン自体を表示しない
+        const name = buttonNames?.[key];
+        if (!name) return null; // Hide button if no name assigned
         return (
           <button
             key={key}
