+++ code/frontend/src/shared/components/Header.tsx
@@ -4,25 +4,13 @@ import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import React from 'react';
-
-
-
+import type { HeaderProps } from '../types/header';
 
 /**
- * ヘッダーコンポーネントのProps
- * @property title - 画面タイトル
- */
-type HeaderProps = {
-  title: string;
-};
-
-
-/**
- * ヘッダーコンポーネント
- * 画面タイトルとロゴを表示。開発環境では色を変更
+ * Header component
+ * Displays screen title and logo. Changes color in development environment.
  */
 const Header: React.FC<HeaderProps> = ({ title }) => {
-  // NODE_ENV で色を切り替え（開発時は赤、リリース時は青）
   const isDev = process.env.NODE_ENV !== 'production';
   return (
     <AppBar position="static" sx={{
@@ -31,9 +19,9 @@ const Header: React.FC<HeaderProps> = ({ title }) => {
     }}>
       <Toolbar>
         <Box sx={{ display: 'flex', alignItems: 'center', flexGrow: 1 }}>
-          {/* ロゴ画像 */}
+          {/* Logo image */}
           <img src="/logo192.png" alt="Logo" style={{ height: 40, marginRight: 16 }} />
-          {/* タイトル */}
+          {/* Title */}
           <Typography variant="h6" component="div" sx={{ fontWeight: 'bold' }}>
             {title}
           </Typography>
