+++ code/frontend/src/shared/types/header.ts
@@ -1,0 +1,9 @@
+/**
+ * Header component types
+ */
+export type HeaderProps = {
+  /**
+   * Screen title
+   */
+  title: string;
+};
