+++ code/frontend/src/shared/types/loading-context.ts
@@ -1,0 +1,7 @@
+/**
+ * LoadingContext types
+ */
+export interface LoadingContextType {
+  loading: boolean;
+  setLoading: (value: boolean) => void;
+}
