+++ code/frontend/src/shared/types/loading.ts
@@ -1,0 +1,7 @@
+/**
+ * LoadingBackdrop component types
+ */
+export type LoadingBackdropProps = {
+  open: boolean;
+  message?: string;
+};
