+++ code/frontend/src/shared/types/qr-reader.ts
@@ -1,0 +1,9 @@
+/**
+ * QrReader component types
+ */
+export type QrReaderProps = {
+  onScan: (result: string) => void;
+  onError?: (error: string) => void;
+  width?: number;
+  height?: number;
+};
